+++ caravan/tests/commands/test_list.py
@@ -5,6 +5,7 @@ import httpretty
 from abduct import captured, out, err
 from freezegun import freeze_time
 
+from caravan.tests.util import mock_args
 from caravan.commands.list import Command
 
 
@@ -43,7 +44,8 @@ def run_command(additional_args):
     args = ['-d', 'DOMAIN']
     args += additional_args
     with captured(out(), err()) as (stdout, stderr):
-        Command.main(args=args)
+        with mock_args(args):
+            Command.main()
 
     return stdout, stderr