+++ CHANGES.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-0.2.1 (unreleased)
+0.2.1 (2015-12-07)
 ------------------
 
 - Add command caravan-domain-list